Migrations add

+++ CleanArchMvc/CleanArchMvc.Domain/Entities/Product.cs
@@ -5,34 +5,34 @@ namespace CleanArchMvc.Domain.Entities
 {
     public sealed class Product : Entity
     {
-       
+
         public string Name { get; private set; }
         public string Description { get; private set; }
-        public decimal Price { get; private set; }
+        public double Price { get; private set; }
         public int Stock { get; private set; }
         public string Image { get; private set; }
 
-        public Product(string name, string description, decimal price, int stock, string image)
+        public Product(string name, string description, double price, int stock, string image)
         {
             ValidateDomain(name, description, price, stock, image);
         }
 
-        public Product(int id, string name, string description, decimal price, int stock, string image)
+        public Product(int id, string name, string description, double price, int stock, string image)
         {
             DomainExceptionValidation.When(id < 0, "Invalid Id value.");
             Id = id;
             ValidateDomain(name, description, price, stock, image);
         }
 
-        public void Update(string name, string description, decimal price, int stock, string image, int categoryId)
+        public void Update(string name, string description, double price, int stock, string image, int categoryId)
         {
             ValidateDomain(name, description, price, stock, image);
             CategoryId = categoryId;
         }
 
-        private void ValidateDomain(string name, string description, decimal price, int stock, string image)
+        private void ValidateDomain(string name, string description, double price, int stock, string image)
         {
-            DomainExceptionValidation.When(string.IsNullOrEmpty(name), 
+            DomainExceptionValidation.When(string.IsNullOrEmpty(name),
                 "Invalid name. Name is required");
 
             DomainExceptionValidation.When(name.Length < 3,
@@ -41,7 +41,7 @@ namespace CleanArchMvc.Domain.Entities
             DomainExceptionValidation.When(string.IsNullOrEmpty(description),
                 "Invalid description. Description is required");
 
-            DomainExceptionValidation.When(description.Length < 5, 
+            DomainExceptionValidation.When(description.Length < 5,
                 "Invalid description, too short, minimum 5 characters");
 
             DomainExceptionValidation.When(price < 0, "Invalid price value");
